Add the Socrates is mortal syllogism
Since Prolog is a logic language, I thought a good problem to tackle as
my first independent project, a simple syllogism.

The source file is

```
% file: socrates.pro
man(socrates).         % Socrates is a man
mortal(X) :- man(X).   % X is mortal if X is man
```

To query this database, open the Prolog shell, `gprolog` and enter
`[socrate].`. To ask if socrates is mortal, `mortal(socrates)`. As
hoped, this returns `yes`.