Map colorings example
This commit is adapted from J.R. Fisher's tutorial (see README for
link).

A map can be expressed as a graph. Each link is expressed as a fact,
e.g. `adj(1,2)` says that regions 1 and 2 are adjacent. Another set of
facts associates each region with a color, e.g. `color(1,red)`. Now a
rule can be written that specifies the conditions of a conflict

```
conflict(X,Y) :- adj(X,Y) , color(X,C) , color(Y,C).
```

Nothing here is particularly new. But I should comment on the a term
J.R. Fisher uses: `unit clause`. According to Wikipedia, a `unit clause`
is a single literal. A single literal is an atomic formula or its
negation. An atomic formula can be a propositional variable (a
propositional variable is a boolean symbol) or in predicate logic a
predicate symbol applied to some terms. Yay for recursive definitions
... anyway, in this case, the atomic formulae are the predicates `adj`
and the terms X and Y.

Also notable, `unit propagation` is the process of reducing a set of
unit clauses to simpler, but equivalent, expressions. This is a super
cool idea which is explained very well on Wikipedia:Unit_propagation.